// src/main/firefox-client/models/DeviceInfo.ts
export interface DeviceInfo {
    appid: string
    apptype: string
    vendor: string
    name: string
    version: string
    appbuildid: string
    platformbuildid: string
    geckobuildid: string
    platformversion: string
    geckoversion: string
    locale: string
    endianness: string
    hostname: string
    os: string
    platform: string
    hardware: string
    arch: string
    processor: string
    compiler: string
    profile: string
    channel: string
    dpi: number
    useragent: string
    width: number
    height: number
    brandName: string
    deviceName: string
    screenshot: string
}